Move laning page factory to separate module
