+++ network-api/networkapi/wagtailpages/tests/research_hub/base.py
@@ -1,3 +1,4 @@
+from networkapi.wagtailpages.factory.research_hub import landing_page as landing_page_factory
 from networkapi.wagtailpages.factory import research_hub as research_factory
 from networkapi.wagtailpages.tests import base as test_base
 
@@ -10,7 +11,7 @@ class ResearchHubTestCase(test_base.WagtailpagesTestCase):
 
     @classmethod
     def _setup_research_hub_structure(cls, homepage):
-        cls.landing_page = research_factory.ResearchLandingPageFactory(
+        cls.landing_page = landing_page_factory.ResearchLandingPageFactory(
             parent=homepage,
         )
         cls.library_page = research_factory.ResearchLibraryPageFactory(
